[openssl] move OpenSSL code to network_openssl.c file
All OpenSSL specific code from connection_glue.c move into
network_openssl.c file.
The new network_openssl.h file is the header of this file to free
network_backends.h of OpenSSL.
