+++ src/connections.h
@@ -21,7 +21,4 @@ handler_t connection_handle_read_post_state(server *srv, connection *con);
 handler_t connection_handle_read_post_error(server *srv, connection *con, int http_status);
 void connection_response_reset(server *srv, connection *con);
 
-int connection_accepted_openssl(server *srv, connection *con);
-int connection_close_openssl(server *srv, connection *con);
-void connection_shutdown_openssl(server *srv, connection *con);
 #endif
