+++ src/network.c
@@ -9,6 +9,7 @@
 #include "configfile.h"
 
 #include "network_backends.h"
+#include "network_openssl.h"
 #include "sys-mmap.h"
 #include "sys-socket.h"
