+++ src/network_backends.h
@@ -75,41 +75,6 @@ int network_write_chunkqueue_writev(server *srv, connection *con, int fd, chunkq
 int network_write_chunkqueue_sendfile(server *srv, connection *con, int fd, chunkqueue *cq, off_t max_bytes); /* fallback to write */
 #endif
 
-#if defined(USE_OPENSSL)
-int network_init_openssl(server *srv);
-void network_free_openssl(server *srv);
-int network_write_chunkqueue_openssl(server *srv, connection *con, void *ssl, chunkqueue *cq, off_t max_bytes);
-void https_cgi_env_openssl(server *srv, connection *con);
-void https_response_prepare_openssl(server *srv, connection *con);
-#else
-# include "log.h"
-static inline int network_init_openssl(server *srv)
-{
-	UNUSED(srv);
-	return 0;
-}
-static inline void network_free_openssl(server *srv)
-{
-	UNUSED(srv);
-}
-static inline int network_write_chunkqueue_openssl(server *srv, connection *con, void *ssl, chunkqueue *cq, off_t max_bytes)
-{
-	UNUSED(srv);UNUSED(con);UNUSED(ssl);UNUSED(cq);UNUSED(max_bytes);
-	log_error_write(srv, __FILE__, __LINE__, "s",
-			"ssl support is missing, recompile with --with-openssl");
-
-	return -1;
-}
-static inline void https_cgi_env_openssl(server *srv, connection *con)
-{
-	UNUSED(srv); UNUSED(con);
-}
-static inline void https_response_prepare_openssl(server *srv, connection *con)
-{
-	UNUSED(srv); UNUSED(con);
-}
-#endif
-
 /* write next chunk(s); finished chunks are removed afterwards after successful writes.
  * return values: similar as backends (0 succes, -1 error, -2 remote close, -3 try again later (EINTR/EAGAIN)) */
 /* next chunk must be MEM_CHUNK. use write()/send() */
