+++ src/network_openssl.c
@@ -1,6 +1,7 @@
 #include "first.h"
 
 #include "network_backends.h"
+#include "network_openssl.h"
 
 #if defined(USE_OPENSSL)
 #  ifndef USE_OPENSSL_KERBEROS
@@ -904,5 +905,227 @@ void https_response_prepare_openssl(server *srv, connection *con) {
 		}
 	}
 	X509_free(xs);
+}
+
+int connection_accepted_openssl(server *srv, connection *con) {
+	specific_config *s = con->srv_socket->conf;
+	openssl_config *sslconf = s->ssl_conf;
+
+	if (NULL == (con->ssl = (void *)SSL_new(sslconf->ssl_ctx))) {
+		log_error_write(srv, __FILE__, __LINE__, "ss", "SSL:",
+				ERR_error_string(ERR_get_error(), NULL));
+		return -1;
+	}
+
+	con->renegotiations = 0;
+	SSL_set_app_data((SSL *)con->ssl, con);
+	SSL_set_accept_state((SSL *)con->ssl);
+#ifndef OPENSSL_NO_TLSEXT
+	if (con->tlsext_server_name)
+		buffer_reset(con->tlsext_server_name);
+	else
+		con->tlsext_server_name = buffer_init();
+#endif
+
+
+	if (1 != (SSL_set_fd((SSL *)con->ssl, con->fd))) {
+		log_error_write(srv, __FILE__, __LINE__, "ss", "SSL:",
+				ERR_error_string(ERR_get_error(), NULL));
+		return -1;
+	}
+	return 0;
+}
+
+int connection_close_openssl(server *srv, connection *con) {
+	UNUSED(srv);
+	if (con->ssl) SSL_free((SSL *)con->ssl);
+	con->ssl = NULL;
+	return 0;
+}
+
+int connection_handle_read_openssl(server *srv, connection *con) {
+	int r, ssl_err, len;
+	char *mem = NULL;
+	size_t mem_len = 0;
+
+	if (!con->srv_socket->is_ssl) return -1;
+
+	ERR_clear_error();
+	do {
+		chunkqueue_get_memory(con->read_queue, &mem, &mem_len, 0, SSL_pending((SSL *)con->ssl));
+#if 0
+		/* overwrite everything with 0 */
+		memset(mem, 0, mem_len);
+#endif
+
+		len = SSL_read((SSL *)con->ssl, mem, mem_len);
+		if (len > 0) {
+			chunkqueue_use_memory(con->read_queue, len);
+			con->bytes_read += len;
+		} else {
+			chunkqueue_use_memory(con->read_queue, 0);
+		}
+
+		if (con->renegotiations > 1 && con->conf.ssl_disable_client_renegotiation) {
+			log_error_write(srv, __FILE__, __LINE__, "s", "SSL: renegotiation initiated by client, killing connection");
+			connection_set_state(srv, con, CON_STATE_ERROR);
+			return -1;
+		}
+	} while (len > 0);
+
+	if (len < 0) {
+		int oerrno = errno;
+		switch ((r = SSL_get_error((SSL *)con->ssl, len))) {
+		case SSL_ERROR_WANT_WRITE:
+			con->is_writable = -1;
+		case SSL_ERROR_WANT_READ:
+			con->is_readable = 0;
+
+			/* the manual says we have to call SSL_read with the same arguments next time.
+			 * we ignore this restriction; no one has complained about it in 1.5 yet, so it probably works anyway.
+			 */
+
+			return 0;
+		case SSL_ERROR_SYSCALL:
+			/**
+			 * man SSL_get_error()
+			 *
+			 * SSL_ERROR_SYSCALL
+			 *   Some I/O error occurred.  The OpenSSL error queue may contain more
+			 *   information on the error.  If the error queue is empty (i.e.
+			 *   ERR_get_error() returns 0), ret can be used to find out more about
+			 *   the error: If ret == 0, an EOF was observed that violates the
+			 *   protocol.  If ret == -1, the underlying BIO reported an I/O error
+			 *   (for socket I/O on Unix systems, consult errno for details).
+			 *
+			 */
+			while((ssl_err = ERR_get_error())) {
+				/* get all errors from the error-queue */
+				log_error_write(srv, __FILE__, __LINE__, "sds", "SSL:",
+						r, ERR_error_string(ssl_err, NULL));
+			}
+
+			switch(oerrno) {
+			default:
+				log_error_write(srv, __FILE__, __LINE__, "sddds", "SSL:",
+						len, r, oerrno,
+						strerror(oerrno));
+				break;
+			}
+
+			break;
+		case SSL_ERROR_ZERO_RETURN:
+			/* clean shutdown on the remote side */
+
+			if (r == 0) {
+				/* FIXME: later */
+			}
+
+			/* fall thourgh */
+		default:
+			while((ssl_err = ERR_get_error())) {
+				switch (ERR_GET_REASON(ssl_err)) {
+				case SSL_R_SSL_HANDSHAKE_FAILURE:
+			      #ifdef SSL_R_TLSV1_ALERT_UNKNOWN_CA
+				case SSL_R_TLSV1_ALERT_UNKNOWN_CA:
+			      #endif
+			      #ifdef SSL_R_SSLV3_ALERT_CERTIFICATE_UNKNOWN
+				case SSL_R_SSLV3_ALERT_CERTIFICATE_UNKNOWN:
+			      #endif
+			      #ifdef SSL_R_SSLV3_ALERT_BAD_CERTIFICATE
+				case SSL_R_SSLV3_ALERT_BAD_CERTIFICATE:
+			      #endif
+					if (!con->conf.log_ssl_noise) continue;
+					break;
+				default:
+					break;
+				}
+				/* get all errors from the error-queue */
+				log_error_write(srv, __FILE__, __LINE__, "sds", "SSL:",
+				                r, ERR_error_string(ssl_err, NULL));
+			}
+			break;
+		}
+
+		connection_set_state(srv, con, CON_STATE_ERROR);
+
+		return -1;
+	} else { /*(len == 0)*/
+		con->is_readable = 0;
+		/* the other end close the connection -> KEEP-ALIVE */
+
+		return -2;
+	}
+}
+
+void connection_shutdown_openssl(server *srv, connection *con) {
+	if (SSL_is_init_finished((SSL *)con->ssl)) {
+		int ret, ssl_r;
+		unsigned long err;
+		ERR_clear_error();
+		switch ((ret = SSL_shutdown((SSL *)con->ssl))) {
+		case 1:
+			/* ok */
+			break;
+		case 0:
+			/* wait for fd-event
+			 *
+			 * FIXME: wait for fdevent and call SSL_shutdown again
+			 *
+			 */
+			ERR_clear_error();
+			if (-1 != (ret = SSL_shutdown((SSL *)con->ssl))) break;
+
+			/* fall through */
+		default:
+
+			switch ((ssl_r = SSL_get_error((SSL *)con->ssl, ret))) {
+			case SSL_ERROR_ZERO_RETURN:
+				break;
+			case SSL_ERROR_WANT_WRITE:
+				/*con->is_writable = -1;*//*(no effect; shutdown() called below)*/
+			case SSL_ERROR_WANT_READ:
+				break;
+			case SSL_ERROR_SYSCALL:
+				/* perhaps we have error waiting in our error-queue */
+				if (0 != (err = ERR_get_error())) {
+					do {
+						log_error_write(srv, __FILE__, __LINE__, "sdds", "SSL:",
+								ssl_r, ret,
+								ERR_error_string(err, NULL));
+					} while((err = ERR_get_error()));
+				} else if (errno != 0) { /* ssl bug (see lighttpd ticket #2213): sometimes errno == 0 */
+					switch(errno) {
+					case EPIPE:
+					case ECONNRESET:
+						break;
+					default:
+						log_error_write(srv, __FILE__, __LINE__, "sddds", "SSL (error):",
+							ssl_r, ret, errno,
+							strerror(errno));
+						break;
+					}
+				}
+
+				break;
+			default:
+				while((err = ERR_get_error())) {
+					log_error_write(srv, __FILE__, __LINE__, "sdds", "SSL:",
+							ssl_r, ret,
+							ERR_error_string(err, NULL));
+				}
+
+				break;
+			}
+		}
+		ERR_clear_error();
+	}
+#ifndef OPENSSL_NO_TLSEXT
+	if (con->tlsext_server_name) {
+		buffer_reset(con->tlsext_server_name);
+		buffer_free(con->tlsext_server_name);
+		con->tlsext_server_name = NULL;
+	}
+#endif
 }
 #endif /* USE_OPENSSL */
